requirements.txt
cairocffi==1.6.1
CairoSVG==2.7.1
cffi==1.16.0
cssselect2==0.7.0
defusedxml==0.7.1
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
pycparser==2.21
tinycss2==1.2.1
tqdm==4.66.2
webencodings==0.5.1
